requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.1.0
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.19
django-filter==23.2
django-import-export==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
MarkupPy==1.14
MarkupSafe==2.1.3
mypy==1.1.1
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
pre-commit==3.4.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.0
pytest==7.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.2
ruff==0.0.287
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.7.1
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
